extract rules for selections
